chore(user-auth): factor out common code
